Melakukan banyak perubahan Ini

+++ src/main/java/id/ac/pcr/springhibernate/model/Mahasiswa.java
@@ -8,7 +8,7 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 
 /**
- * Created with IntelliJ IDEA.
+ * Created with LongelliJ IDEA.
  * User: isaninside
  * Date: 4/3/13
  * Time: 10:10 AM
@@ -23,17 +23,17 @@ public class Mahasiswa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mahasiswa_id")
-    private int id;
+    private Long id;
     @Column(name = "nama")
     private String nama;
     @Column(name = "nim")
     private String nim;
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
